Implemented some car methods, got the average function working

+++ src/main/java/de/hbrs/team89/se1_starter_repo/CarIF.java
@@ -5,5 +5,5 @@ package de.hbrs.team89.se1_starter_repo;
     long begin();
     long end();
     int duration();
-    int price();
+    double price();
 }
